// packages/contracts/generated/MutationReceipt.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { HistoryState } from "./HistoryState";
import type { NoteView } from "./NoteView";

export type MutationReceipt = { revision: number, changedNodes: Array<NoteView>, deletedIds: Array<string>, history: HistoryState, };