// packages/contracts/generated/ViewportPage.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { NoteView } from "./NoteView";

export type ViewportPage = { pageId: string, anchorId: string | null, beforeCursor: string | null, afterCursor: string | null, nodes: Array<NoteView>, };